added academy gateway skeleton. Moved gateway interfaces into separate folder
