+++ SingleViewApi.Tests/V1/UseCase/GetAllNotesUseCaseTests.cs
@@ -11,6 +11,7 @@ using NUnit.Framework;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.UseCase.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
