+++ SingleViewApi.Tests/V1/UseCase/GetAllUseCaseTests.cs
@@ -9,6 +9,7 @@ using FluentAssertions;
 using Hackney.Core.Testing.Shared;
 using Moq;
 using NUnit.Framework;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
 {
