+++ SingleViewApi.Tests/V1/UseCase/GetCustomerByIdUseCaseTests.cs
@@ -15,6 +15,7 @@ using NUnit.Framework;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Boundary.Response;
 using SingleViewApi.V1.Domain;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.UseCase.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
