+++ SingleViewApi.Tests/V1/UseCase/GetJigsawAuthTokenUseCaseTests.cs
@@ -6,6 +6,7 @@ using ServiceStack;
 using ServiceStack.Redis;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Helpers;
 using SingleViewApi.V1.Helpers.Interfaces;
 using SingleViewApi.V1.UseCase;
