+++ SingleViewApi.Tests/V1/UseCase/GetJigsawNotesUseCaseTests.cs
@@ -10,6 +10,7 @@ using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Boundary.Response;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.UseCase.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
