+++ SingleViewApi.Tests/V1/UseCase/GetNotesUseCaseTests.cs
@@ -9,6 +9,7 @@ using NUnit.Framework;
 using SingleViewApi.V1.Boundary.Response;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
 {
