+++ SingleViewApi.Tests/V1/UseCase/GetPersonApiByIdUseCaseTests.cs
@@ -12,6 +12,7 @@ using NUnit.Framework;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Boundary.Response;
 using SingleViewApi.V1.Domain;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.Tests.V1.UseCase
 {
