+++ SingleViewApi.Tests/V1/UseCase/SearchSingleViewUseCaseTests.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.UseCase;
 
 namespace SingleViewApi.Tests.V1.UseCase
