+++ SingleViewApi.Tests/V1/UseCase/StoreJigsawCredentialsUseCaseTests.cs
@@ -8,6 +8,7 @@ using NUnit.Framework;
 using ServiceStack.Redis;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Helpers;
 using SingleViewApi.V1.Helpers.Interfaces;
 using SingleViewApi.V1.UseCase;
