+++ SingleViewApi/V1/Gateways/CustomerDataSourceGateway.cs
@@ -1,6 +1,7 @@
 using System;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Factories;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Infrastructure;
 
 namespace SingleViewApi.V1.Gateways
