+++ SingleViewApi/V1/Gateways/DataSourceGateway.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using ServiceStack;
 using SingleViewApi.V1.Domain;
 using SingleViewApi.V1.Factories;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Infrastructure;
 
 namespace SingleViewApi.V1.Gateways
