+++ SingleViewApi/V1/Gateways/HousingSearchGateway.cs
@@ -7,6 +7,7 @@ using Hackney.Shared.Person;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
 using SingleViewApi.V1.Boundary.Response;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.V1.Gateways
 {
