+++ SingleViewApi/V1/Gateways/Interfaces/IAcademyGateway.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace SingleViewApi.V1.Gateways.Interfaces;
+
+public interface IAcademyGateway
+{
+    Task<dynamic> GetCouncilTaxAccountsByCustomerName(string firstName, string lastName);
+}
