+++ SingleViewApi/V1/Gateways/Interfaces/ICustomerDataSourceGateway.cs
@@ -1,7 +1,7 @@
 using System;
 using SingleViewApi.V1.Domain;
 
-namespace SingleViewApi.V1.Gateways
+namespace SingleViewApi.V1.Gateways.Interfaces
 {
     public interface ICustomerDataSourceGateway
     {
