+++ SingleViewApi/V1/Gateways/Interfaces/IDataSourceGateway.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using SingleViewApi.V1.Domain;
 
-namespace SingleViewApi.V1.Gateways
+namespace SingleViewApi.V1.Gateways.Interfaces
 {
     public interface IDataSourceGateway
     {
