+++ SingleViewApi/V1/Gateways/Interfaces/IEqualityInformationGateway.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using SingleViewApi.V1.Boundary.Response;
 
-namespace SingleViewApi.V1.Gateways;
+namespace SingleViewApi.V1.Gateways.Interfaces;
 
 public interface IEqualityInformationGateway
 {
