+++ SingleViewApi/V1/Gateways/Interfaces/IHousingSearchGateway.cs
@@ -1,9 +1,7 @@
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using Hackney.Shared.Person;
 using SingleViewApi.V1.Boundary.Response;
 
-namespace SingleViewApi.V1.Gateways
+namespace SingleViewApi.V1.Gateways.Interfaces
 {
     public interface IHousingSearchGateway
     {
