+++ SingleViewApi/V1/Gateways/Interfaces/INotesGateway.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using SingleViewApi.V1.Boundary.Request;
 using SingleViewApi.V1.Boundary.Response;
 
-namespace SingleViewApi.V1.Gateways
+namespace SingleViewApi.V1.Gateways.Interfaces
 {
     public interface INotesGateway
     {
