+++ SingleViewApi/V1/UseCase/CreateNoteUseCase.cs
@@ -4,6 +4,7 @@ using SingleViewApi.V1.UseCase.Interfaces;
 using Hackney.Core.Logging;
 using SingleViewApi.V1.Boundary.Request;
 using SingleViewApi.V1.Boundary.Response;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.V1.UseCase
 {
