+++ SingleViewApi/V1/UseCase/GetJigsawAuthTokenUseCase.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Hackney.Core.Logging;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Helpers;
 using SingleViewApi.V1.Helpers.Interfaces;
 using SingleViewApi.V1.UseCase.Interfaces;
