+++ SingleViewApi/V1/UseCase/GetPersonAPIByIdUseCase.cs
@@ -10,6 +10,7 @@ using Hackney.Shared.ContactDetail.Domain;
 using Hackney.Shared.Person;
 using SingleViewApi.V1.Boundary;
 using SingleViewApi.V1.Domain;
+using SingleViewApi.V1.Gateways.Interfaces;
 
 namespace SingleViewApi.V1.UseCase
 {
