+++ SingleViewApi/V1/UseCase/StoreJigsawCredentialsUseCase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading.Tasks;
 using Hackney.Core.Logging;
 using SingleViewApi.V1.Gateways;
+using SingleViewApi.V1.Gateways.Interfaces;
 using SingleViewApi.V1.Helpers.Interfaces;
 using SingleViewApi.V1.UseCase.Interfaces;
